fix(di): switch to @Provides for use cases to resolve injection errors.
