+++ core/di/src/main/java/com/moviescatalog/core/di/NetworkModule.kt
@@ -1,6 +1,7 @@
 package com.moviescatalog.core.di
 
 
+
 import com.moviescatalog.core.util.Constants
 import com.moviescatalog.data.remote.api.MovieApiService
 import dagger.Module
