+++ core/di/src/main/java/com/moviescatalog/core/di/RepositoryModule.kt
@@ -7,14 +7,12 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
-
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RepositoryModule {
 
     @Binds
-    @Singleton
     abstract fun bindMovieRepository(
-        movieRepositoryImpl: MovieRepositoryImpl
+        impl: MovieRepositoryImpl
     ): MovieRepository
 }
