+++ core/network/build.gradle.kts
@@ -1,12 +1,14 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
 }
 
 android {
     namespace = "com.moviescatalog.core.network"
     compileSdk = libs.versions.compileSdk.get().toInt()
-
+    defaultConfig {
+        minSdk = libs.versions.minSdk.get().toInt()
+    }
     buildTypes {
         release {
             isMinifyEnabled = false
