+++ data/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.hilt)
     alias(libs.plugins.google.devtools.ksp)
@@ -9,7 +9,9 @@ android {
     namespace = "com.moviescatalog.data"
     compileSdk = libs.versions.compileSdk.get().toInt()
 
-
+    defaultConfig {
+        minSdk = libs.versions.minSdk.get().toInt()
+    }
     buildTypes {
         release {
             isMinifyEnabled = false
