+++ domain/src/main/java/com/moviescatalog/domain/usecase/GetPopularMoviesUseCase.kt
@@ -7,11 +7,3 @@ import javax.inject.Inject
 interface GetPopularMoviesUseCase {
     suspend operator fun invoke(page: Int): Result<List<Movie>>
 }
-class GetPopularMoviesUseCaseImpl @Inject constructor(
-    private val repository: MovieRepository
-) : GetPopularMoviesUseCase {
-
-    override suspend fun invoke(page: Int): Result<List<Movie>> {
-        return repository.getMovies("popularity.desc", page)
-    }
-}
