+++ domain/src/main/java/com/moviescatalog/domain/usecase/GetTopRevenueMoviesUseCase.kt
@@ -8,10 +8,3 @@ import javax.inject.Inject
 interface GetTopRevenueMoviesUseCase {
     suspend operator fun invoke(page: Int): Result<List<Movie>>
 }
-class GetTopRevenueMoviesUseCaseImpl @Inject constructor(
-    private val repository: MovieRepository
-) : GetTopRatedMoviesUseCase {
-    override suspend fun invoke(page: Int): Result<List<Movie>>{
-        return repository.getMovies("vote_average.desc", page)
-    }
-}
